fix(stream error propagation): deactivates dependecy streams on error only if deactivation guard pas
when stream errors it should not cause it's dependencies to error if they have other dependents or
have subscribers.  Tests coming. This actually makes notifying  up the chain on complete and on
error  the same - will refactor in the next commit. Also tests are coming.

BREAKING CHANGE: deactivates dependecy streams on error only if deactivation guard passes